@@ -7,6 +7,7 @@ using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using NLog.Web;
 
 namespace FarmerAPI
 {
@@ -19,6 +20,9 @@ namespace FarmerAPI
 
         public static IWebHostBuilder BuildWebHost(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+                .UseStartup<Startup>().ConfigureLogging(logging=> {
+					logging.ClearProviders();
+					logging.SetMinimumLevel(LogLevel.Trace);
+				}).UseNLog();
     }
 }
